Created team display with set starter slots
Problem:
User could have multiple players at a position as starters on the lineup at the same time. This would break the optimization features of the program because it would just add all of the players to the starting lineup to maximize points.

Fix:
Created a team display that limits the amount of players in the starting lineup and displays a label 'Please add a player' when there is no starter at that position.

Closes #6